requirements.txt
requests==2.31.0
python-dotenv==1.0.1
pandas
pandas-ta
ccxt==4.2.14
schedule
